update turboquant offload
