+++ configs/lingbot_fast/kv_calib/lingbot_fast_i2v_kv_turboquant_calib.json
@@ -4,7 +4,7 @@
     "text_len": 512,
     "target_height": 720,
     "target_width": 1280,
-    "self_attn_1_type": "sage_attn2_k_int8_v_fp8",
+    "self_attn_1_type": "sage_attn2",
     "cross_attn_1_type": "sage_attn2",
     "cross_attn_2_type": "sage_attn2",
     "sample_guide_scale": 1.0,
@@ -22,11 +22,11 @@
         "timesteps_index": [0, 179, 358, 679],
         "sink_size": 3,
         "kv_quant": {
-            "calibrate": false,
-            "calib_path": "calib_kv.pt",
-            "quant_scheme": "sage",
-            "k_cache_type": "int8",
-            "v_cache_type": "fp8"
+            "calibrate": true,
+            "quant_scheme": "turboquant",
+            "codebook_dir": "/data/nvme4/gushiqiao/new/LightX2V/save_results/turboquant_codebooks",
+            "key_bits": 4,
+            "value_bits": 2
         },
         "kv_offload": false
     },
