+++ configs/lingbot_fast/lingbot_fast_i2v_kv_turboquant.json
@@ -4,17 +4,18 @@
     "text_len": 512,
     "target_height": 720,
     "target_width": 1280,
-    "self_attn_1_type": "sage_attn3",
+    "self_attn_1_type": "sage_attn2",
     "cross_attn_1_type": "sage_attn2",
     "cross_attn_2_type": "sage_attn2",
     "sample_guide_scale": 1.0,
     "sample_shift": 10.0,
     "enable_cfg": false,
-    "cpu_offload": true,
+    "cpu_offload": false,
     "offload_granularity": "block",
-    "t5_cpu_offload": true,
-    "vae_cpu_offload": true,
+    "t5_cpu_offload": false,
+    "vae_cpu_offload": false,
     "use_image_encoder": false,
+    "dit_original_ckpt": "/data/nvme4/models/lingbot-world-base-cam/lingbot_world_fast/",
     "ar_config": {
         "local_attn_size": 21,
         "num_frame_per_chunk": 3,
@@ -22,16 +23,11 @@
         "sink_size": 3,
         "kv_quant": {
             "quant_scheme": "turboquant",
-            "calibrate": true,
-            "calib_path": "save_results/my_kv_calib.pt",
-            "turboquant_codebook_out_dir": "save_results/turboquant_codebooks",
+            "calibrate": false,
+            "codebook_dir": "/data/nvme4/gushiqiao/new/LightX2V/save_results/turboquant_codebooks",
             "key_bits": 4,
-            "value_bits": 2,
-            "use_qjl": false,
-            "turboquant_seed": 42,
-            "per_layer_compressors": true,
-            "value_quant_mode": "mse"
-          },
+            "value_bits": 2
+        },
         "kv_offload": true
     },
     "rms_norm_type": "self_forcing"
