+++ scripts/lingbot/run_lingbot_fast_i2v.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 
 # set path firstly
-lightx2v_path=/mnt/devsft_afs_2/gushiqiao/LightX2V
-model_path=/models/lingbot-world-fast/
+lightx2v_path=/data/nvme4/gushiqiao/new/LightX2V
+model_path=/data/nvme4/models/lingbot-world-base-cam
 
-export CUDA_VISIBLE_DEVICES=0
+export CUDA_VISIBLE_DEVICES=7
 
 # set environment variables
 source ${lightx2v_path}/scripts/base/base.sh
@@ -13,9 +13,9 @@ python -m lightx2v.infer \
 --model_cls lingbot_world_fast \
 --task i2v \
 --model_path $model_path \
---config_json ${lightx2v_path}/configs/lingbot_fast/lingbot_fast_i2v.json \
+--config_json /data/nvme4/gushiqiao/new/LightX2V/configs/lingbot_fast/lingbot_fast_i2v_kv_turboquant.json \
 --prompt "A serene lakeside scene with a lone tree standing in calm water, surrounded by distant snow-capped mountains under a bright blue sky with drifting white clouds — gentle ripples reflect the tree and sky, creating a tranquil, meditative atmosphere." \
 --negative_prompt "" \
---image_path /mnt/devsft_afs_2/gushiqiao/lingbot-world/examples/03/image.jpg \
---action_path /mnt/devsft_afs_2/gushiqiao/lingbot-world/examples/03/ \
+--image_path /data/nvme4/gushiqiao/lingbot-world/examples/03/image.jpg \
+--action_path /data/nvme4/gushiqiao/lingbot-world/examples/03/ \
 --save_result_path ${lightx2v_path}/save_results/output_lightx2v_lingbot_fast_i2v.mp4
